Adds 14.7 solution; improves documentation
Exercise 14.7 added a quarter and chocolate to the finite state machine.
This commit completes the exercise but also shows how much needed to be
changed when the quarter and chocolate were added. I think there's a
better structure for adding a new state/node and its new
transitions/arcs. One option would be to overwrite the arc with a new
arc when its from and to states/nodes matched. That works but needs
immutability added to it. It might also entail adding nodes and arcs to
the end of the program file, which would make it less readable. It would
be a good time to learn tests or create testing in comments to check
that all transitions remain the same or change in the desired ways.